refactor(runtime): fold @maka/code-mode into the runtime workspace
With the SDK-superseded scaffolding gone, the package was 224 lines of
integration glue over `@ai-sdk/code-mode` — the product execution
policy, the result shapes the backend publishes, and the adapter that
bridges sandbox tool calls onto host tools and waits for them to drain.
At that size a standalone workspace costs more than it explains: its
only consumer is `AiSdkBackend`, which now imports it as a sibling.

`index.ts` and `quickjs.ts` merge into `packages/runtime/src/code-mode.ts`,
since the split only existed to give the package a public entry point
distinct from its implementation. Its tests move alongside.

The workspace manifest, tsconfig and project reference, both build
chains, the desktop and CLI build chains, the stale-dist pairs, the
release script's internal-package and build-order lists, the CLI
validation path filters, and the Windows sandbox build step all drop
their entry. `@ai-sdk/code-mode` moves from the deleted package's
dependencies to the runtime's, so the release script's `npm ls` closure
still reaches it; both third-party notice inventories already listed it
under its own name and are unchanged.

The dependabot ignore for typescript majors kept a comment naming the
deleted path. Its subject — an in-house transpiler on the typescript JS
API pinned to 5.9 — no longer exists anywhere in the repo, and no
workspace pins typescript. The comment now says so; whether to drop the
ignore is a dependency-policy call left to a maintainer.

Generated-by: Claude Code
