+++ .github/dependabot.yml
@@ -12,10 +12,12 @@ updates:
         update-types:
           - minor
           - patch
-    # packages/code-mode depends on the typescript JS API (transpileModule,
-    # ScriptTarget, …) at runtime; the 6.0+ native compiler removed those
-    # exports, so the package stays pinned to 5.9. The workspace root already
-    # compiles on 7.x, so ignoring >=6 only affects that runtime pin.
+    # Retained from when an in-house Code Mode transpiler used the typescript
+    # JS API (transpileModule, ScriptTarget, …) at runtime and had to stay on
+    # 5.9, since the 6.0+ native compiler removed those exports. That
+    # transpiler is gone — @ai-sdk/code-mode owns transpilation now, and no
+    # workspace pins typescript — so this ignore no longer has a subject.
+    # Dropping it is a dependency-policy call, not this refactor's to make.
     ignore:
       - dependency-name: 'typescript'
         versions:
