+++ packages/runtime/package.json
@@ -125,13 +125,13 @@
   },
   "dependencies": {
     "@ai-sdk/anthropic": "4.0.39",
+    "@ai-sdk/code-mode": "1.0.23",
     "@ai-sdk/cohere": "4.0.27",
     "@ai-sdk/google": "4.0.44",
     "@ai-sdk/open-responses": "2.0.28",
     "@ai-sdk/openai": "4.0.42",
     "@ai-sdk/openai-compatible": "3.0.30",
     "@larksuiteoapi/node-sdk": "1.72.0",
-    "@maka/code-mode": "0.1.0",
     "@maka/core": "0.1.0",
     "@modelcontextprotocol/sdk": "1.30.0",
     "@mozilla/readability": "^0.6.0",
