+++ scripts/check-stale-dist.mjs
@@ -27,7 +27,6 @@ const here = dirname(fileURLToPath(import.meta.url));
 const repoRoot = resolve(here, '..');
 
 const pairs = [
-  { pkg: '@maka/code-mode', src: 'packages/code-mode/src', dist: 'packages/code-mode/dist' },
   { pkg: '@maka/core', src: 'packages/core/src', dist: 'packages/core/dist' },
   { pkg: '@maka/storage', src: 'packages/storage/src', dist: 'packages/storage/dist' },
   { pkg: '@maka/runtime', src: 'packages/runtime/src', dist: 'packages/runtime/dist' },
