+++ scripts/ci-test-plan.mjs
@@ -42,7 +42,6 @@ const CLI_PACKAGE_FILES = new Set([
 
 const CLI_PACKAGE_WORKSPACES = [
   'packages/cli',
-  'packages/code-mode',
   'packages/core',
   'packages/eval',
   'packages/mcp',
